requirements.txt
Flask==3.0.2
celery==5.3.6
redis==5.0.1
openai==1.3.0
google-cloud-translate==3.20.2
elevenlabs==0.2.24
moviepy==1.0.3
numpy>=2.1.3
opencv-python==4.9.0.80
torch==2.6.0
torchaudio==2.6.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
pytube==15.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
google-cloud-texttospeech==2.15.0
google-cloud-storage==2.14.0
scipy==1.12.0 